ProcessResults/requirements.txt
requests
openpyxl

mysql-connector-python
python-dotenv
